+++ scripts/tracking.js
@@ -171,7 +171,9 @@ async function track(chapter = false, volume = false, onlySite = false, reduceIf
 							sitesScripts[site.key].track({
 								id: site.tracking.id,
 								chapters: vars.chapter,
+								chaptersInt: vars.chapters ? Math.floor(vars.chapters) : vars.chapters,
 								volumes: vars.volume,
+								volumesInt: vars.volume ? Math.floor(vars.volume) : vars.volume,
 								mainPath: mainPath,
 							});
 						}